more mac build fixes

+++ onnxruntime/contrib_ops/cpu/attnlstm/deep_cpu_attn_lstm.h
@@ -97,7 +97,7 @@ class DeepCpuAttnLstmOp final : public OpKernel {
 // The alternative would be to create a threadpool in each call to Compute but that would incur thread creation
 // cost on every call.
 #ifdef USE_EIGEN_THREADPOOL
-  mutable Eigen::NonBlockingThreadPool ttp_{std::thread::hardware_concurrency()};
+  mutable Eigen::NonBlockingThreadPool ttp_{static_cast<int>(std::thread::hardware_concurrency())};
 #else
   mutable TaskThreadPool ttp_{std::thread::hardware_concurrency()};
 #endif
